# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  github_com_yixian-huang_inkless_backend_internal_model.JSONMap:
    additionalProperties: true
    type: object
...
  internal_handler_content.GetVersionDetailResponse:
    properties:
      config:
        $ref: '#/definitions/github_com_yixian-huang_inkless_backend_internal_model.JSONMap'
      createdBy:
        type: integer
      id:
        type: integer
      pageKey:
        type: string
      publishedAt:
        type: string
      version:
        type: integer
    type: object
  internal_handler_content.GetVersionsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handler_content.VersionListItem'
        type: array
      total:
        type: integer
    type: object
...
  internal_handler_content.VersionListItem:
    properties:
      changeNote:
        type: string
      createdAt:
        type: string
      operator:
        type: string
      version:
        type: integer
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: Extensible Inkless CMS API for content, themes, media, plugins, and
    site management.
  title: Inkless CMS API
  version: "1.0"
paths:
...
  /admin/content/{pageKey}/versions:
    get:
      parameters:
      - description: Page key
        in: path
        name: pageKey
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.GetVersionsResponse'
      security:
      - BearerAuth: []
      summary: List theme content versions
      tags:
      - Content (Admin)
  /admin/content/{pageKey}/versions/{version}:
    get:
      parameters:
      - description: Page key
        in: path
        name: pageKey
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.GetVersionDetailResponse'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: object
            type: object
      security:
      - BearerAuth: []
      summary: Get theme content version detail
      tags:
      - Content (Admin)
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" for JWT authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"